+++ gasstation/gasstattion.go
@@ -13,14 +13,12 @@ import (
 
 	"github.com/iotexproject/go-pkgs/hash"
 	"github.com/iotexproject/iotex-address/address"
-	"github.com/pkg/errors"
 
 	"github.com/iotexproject/iotex-core/action"
 	"github.com/iotexproject/iotex-core/action/protocol/execution/evm"
 	"github.com/iotexproject/iotex-core/blockchain"
 	"github.com/iotexproject/iotex-core/blockchain/block"
 	"github.com/iotexproject/iotex-core/config"
-	"github.com/iotexproject/iotex-proto/golang/iotextypes"
 )
 
 // BlockDAO represents the block data access object
@@ -105,35 +103,6 @@ func (gs *GasStation) SuggestGasPrice() (uint64, error) {
 		gasPrice = gs.cfg.GasStation.DefaultGas
 	}
 	return gasPrice, nil
-}
-
-// EstimateGasForAction estimate gas for action
-func (gs *GasStation) EstimateGasForAction(actPb *iotextypes.Action) (uint64, error) {
-	selp, err := (&action.Deserializer{}).ActionToSealedEnvelope(actPb)
-	if err != nil {
-		return 0, err
-	}
-	// Special handling for executions
-	if sc, ok := selp.Action().(*action.Execution); ok {
-		callerAddr := selp.SrcPubkey().Address()
-		if callerAddr == nil {
-			return 0, errors.New("failed to get address")
-		}
-		ctx, err := gs.bc.Context(context.Background())
-		if err != nil {
-			return 0, err
-		}
-		_, receipt, err := gs.simulator(ctx, callerAddr, sc, gs.dao.GetBlockHash)
-		if err != nil {
-			return 0, err
-		}
-		return receipt.GasConsumed, nil
-	}
-	gas, err := selp.IntrinsicGas()
-	if err != nil {
-		return 0, err
-	}
-	return gas, nil
 }
 
 type bigIntArray []*big.Int
